Rename: 좀더 적절한 이름으로 컴포넌트명 변경
