--- src/routes/ManageAds/AdsHeader/AdsCreateButton/AdsCreateForm/adsCreateForm.module.scss
@@ -1,19 +1,0 @@
-@use 'styles/constants/colors';
-@use 'styles/mixins/flexbox';
-
-.createFormWrapper {
-  @include flexbox.flexbox(center, center);
-  flex-direction: column;
-
-  ul {
-    width: 100%;
-    margin: 20px 0;
-  }
-
-  button {
-    width: 100px;
-    height: 40px;
-    border: 1px solid colors.$GRAY100;
-    border-radius: 10px;
-  }
-}
